requirements.txt
fastapi[standard]==0.115.0
redis==5.1.1
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
prometheus-client==0.21.0
structlog==24.4.0
tenacity==9.0.0
python-dateutil==2.8.2
